feat: update StockService to return BigDecimal stock prices

+++ chapter-06/src/main/java/expert/os/books/architecture/patterns/remote/StockQueryDto.java
@@ -1,4 +1,6 @@
 package expert.os.books.architecture.patterns.remote;
 
-public record StockQueryDto(double minPrice, double maxPrice, String sector) {
+import java.math.BigDecimal;
+
+public record StockQueryDto(BigDecimal minPrice, BigDecimal maxPrice, String sector) {
 }
